+++ app/src/main/java/com/example/animalcrossing/data/repository/IslandRepository.kt
@@ -1,6 +1,10 @@
 package com.example.animalcrossing.data.repository
 
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.widget.Toast
 import com.example.animalcrossing.data.firebase.AcnhFirebaseRepository
 import com.example.animalcrossing.data.db.IslandDBRepository
 import com.example.animalcrossing.data.db.IslandEntity
@@ -16,7 +20,8 @@ import javax.inject.Singleton
 @Singleton
 class IslandRepository @Inject constructor(
     private val dbRepository: IslandDBRepository,
-    private val apiRepository: AcnhFirebaseRepository
+    private val apiRepository: AcnhFirebaseRepository,
+    private val context: Context
 ) {
     val island: Flow<Island?> = dbRepository.island
         .map { it.asIsland() }
@@ -25,19 +30,32 @@ class IslandRepository @Inject constructor(
 
 
     suspend fun addIsland(name: String, hemisphere: String) {
-        val newIsland = IslandEntity(name = name, hemisphere = hemisphere)
-        apiRepository.createIsland(name, hemisphere)
-        dbRepository.insert(newIsland)
-    }
-    suspend fun deleteIsland(id:Long) {
-        apiRepository.deleteIsland()
-        dbRepository.delete(id)
-        dbRepository.deleteAllLoans()
+        if (isOnline()) {
+            val newIsland = IslandEntity(name = name, hemisphere = hemisphere)
+            apiRepository.createIsland(name, hemisphere)
+            dbRepository.insert(newIsland)
+        } else {
+            showNoInternetToast()
+        }
     }
 
-    suspend fun renameIsland(id:Long, name: String) {
-        apiRepository.renameIsland(name)
-        dbRepository.rename(id, name)
+    suspend fun deleteIsland(id: Long) {
+        if (isOnline()) {
+            apiRepository.deleteIsland()
+            dbRepository.delete(id)
+            dbRepository.deleteAllLoans()
+        } else {
+            showNoInternetToast()
+        }
+    }
+
+    suspend fun renameIsland(id: Long, name: String) {
+        if (isOnline()) {
+            apiRepository.renameIsland(name)
+            dbRepository.rename(id, name)
+        } else {
+            showNoInternetToast()
+        }
     }
 
     suspend fun searchVillagers(query: String): Flow<List<VillagerEntity>> {
@@ -45,13 +63,38 @@ class IslandRepository @Inject constructor(
     }
 
     suspend fun addVillagerToIsland(name: String, islandId: Long) {
-        apiRepository.addVillagerToIsland(name)
-        return dbRepository.addVillagerToIsland(name, islandId)
+        if (isOnline()) {
+            apiRepository.addVillagerToIsland(name)
+            dbRepository.addVillagerToIsland(name, islandId)
+        } else {
+            showNoInternetToast()
+        }
     }
 
     suspend fun deleteVillagerFromIsland(name: String, islandId: Long) {
-        apiRepository.deleteVillagerFromIsland(name)
-        return dbRepository.deleteVillagerFromIsland(name, islandId)
+        if (isOnline()) {
+            apiRepository.deleteVillagerFromIsland(name)
+            dbRepository.deleteVillagerFromIsland(name, islandId)
+        } else {
+            showNoInternetToast()
+        }
+    }
+
+    private fun isOnline(): Boolean {
+        val connectivityManager =
+            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val network = connectivityManager.activeNetwork
+        val networkCapabilities = connectivityManager.getNetworkCapabilities(network)
+        return networkCapabilities != null &&
+                networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+    }
+
+    private fun showNoInternetToast() {
+        Toast.makeText(
+            context,
+            "No hay conexión a Internet. Por favor, comprueba tu conexión e inténtalo de nuevo.",
+            Toast.LENGTH_SHORT
+        ).show()
     }
 
 }
