+++ app/src/main/java/com/example/animalcrossing/data/repository/LoanRepository.kt
@@ -1,6 +1,10 @@
 package com.example.animalcrossing.data.repository
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
 import android.util.Log
+import android.widget.Toast
 import com.example.animalcrossing.data.firebase.AcnhFirebaseRepository
 import com.example.animalcrossing.data.db.LoansDBRepository
 import com.example.animalcrossing.data.db.LoansEntity
@@ -13,21 +17,26 @@ import javax.inject.Singleton
 @Singleton
 class LoanRepository @Inject constructor(
     private val dbRepository: LoansDBRepository,
-    private val apiRepository: AcnhFirebaseRepository
+    private val apiRepository: AcnhFirebaseRepository,
+    private val context: Context
 ) {
 
     val loans: Flow<List<Loan>>
         get() {
             val list = dbRepository.allLoans.map { it.asLoan() }
-
             return list
         }
 
     suspend fun addLoan(title: String, type: String, amountPaid: Int, amountTotal: Int, completed: Boolean): Long {
-        val newLoan = LoansEntity(firebaseId = "", title = title, type = type, amountPaid = amountPaid, amountTotal = amountTotal, completed = completed)
-        val firebaseId = apiRepository.createLoan(newLoan)
-        newLoan.firebaseId = firebaseId
-        return dbRepository.insert(newLoan)
+        if (isOnline()) {
+            val newLoan = LoansEntity(firebaseId = "", title = title, type = type, amountPaid = amountPaid, amountTotal = amountTotal, completed = completed)
+            val firebaseId = apiRepository.createLoan(newLoan)
+            newLoan.firebaseId = firebaseId
+            return dbRepository.insert(newLoan)
+        } else {
+            showNoInternetToast()
+            return -1
+        }
     }
 
     suspend fun getLoan(firebaseId: String): Flow<LoansEntity> {
@@ -35,12 +44,37 @@ class LoanRepository @Inject constructor(
     }
 
     suspend fun deleteLoan(firebaseId: String) {
-        apiRepository.deleteLoan(firebaseId)
-        dbRepository.deleteLoan(firebaseId)
+        if (isOnline()) {
+            apiRepository.deleteLoan(firebaseId)
+            dbRepository.deleteLoan(firebaseId)
+        } else {
+            showNoInternetToast()
+        }
     }
 
     suspend fun updateLoan(loan: Loan) {
-        apiRepository.editLoan(loan);
-        dbRepository.updateLoan(loan)
+        if (isOnline()) {
+            apiRepository.editLoan(loan)
+            dbRepository.updateLoan(loan)
+        } else {
+            showNoInternetToast()
+        }
+    }
+
+    private fun isOnline(): Boolean {
+        val connectivityManager =
+            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val network = connectivityManager.activeNetwork
+        val networkCapabilities = connectivityManager.getNetworkCapabilities(network)
+        return networkCapabilities != null &&
+                networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+    }
+
+    private fun showNoInternetToast() {
+        Toast.makeText(
+            context,
+            "No hay conexión a Internet. Por favor, comprueba tu conexión e inténtalo de nuevo.",
+            Toast.LENGTH_SHORT
+        ).show()
     }
 }
